Module Constraints.Make

Parameters

module Algebra : Module_types.Algebra

Signature

module Label := Algebra.Types.Label

module Type_var := Algebra.Types.Var

module Type_former := Algebra.Types.Former

module Decoded : Module_types.Decoded with type label := Label.t and type 'a former := 'a Type_former.t

Constraints require an explicit term algebra for types.

Which we define by taking the fixpoint of Type_former.t with constraint variables.

type variable = private Constraints__.Import.Int.t

variable is the for the constraint variables

A constraint of type 'a Constraint.t represents a constraint of type 'a.

To acquire the a constraint, we first specify it's term variables and types using an Algebra.

type 'a t

and binding = Algebra.Term_var.t * variable

and def_binding = binding

and 'a let_binding

and 'a let_rec_binding

val sexp_of_t : 'a t -> Constraints__.Import.Sexp.t

fresh () creates a fresh constraint variable.

val fresh : Constraints__.Import.Unit.t -> variable

The module Type provides the concrete representation of types (using constraint type variables) in constraints.

It is the free monad of the functor Type_former.t.

History: This representation was initially used in constraints t, however, the refactor for "Sharing" now uses Shallow_type.t. We however, still use Type for a rich interface.

module Type : sig ... end

The module Shallow_type provides the shallow type definition used within constraints.

This encoding is required for "(Explicit) sharing" of types within constraints.

Types from Type are often referred to as "deep" types.

module Shallow_type : sig ... end

Constraints form a applicative functor, allowing us to combine many constraints into a single one using let%map:

val pat : Typedtree.pattern Constraint.t
val exp : Typedtree.expression Constraint.t

let%map pat and exp in
Texp_fun (pat, ..., exp)

val return : 'a -> 'a t

val map : 'a t -> f:('a -> 'b) -> 'b t

val both : 'a t -> 'b t -> ('a * 'b) t

val (<*>) : ('a -> 'b) t -> 'a t -> 'b t

val (<*) : 'a t -> unit t -> 'a t

val (*>) : unit t -> 'a t -> 'a t

val (>>|) : 'a t -> ('a -> 'b) -> 'b t

val apply : ('a -> 'b) t -> 'a t -> 'b t

val map2 : 'a t -> 'b t -> f:('a -> 'b -> 'c) -> 'c t

val map3 : 'a t -> 'b t -> 'c t -> f:('a -> 'b -> 'c -> 'd) -> 'd t

val all : 'a t list -> 'a list t

val all_unit : unit t list -> unit t

module Applicative_infix : sig ... end

module Open_on_rhs_intf : sig ... end

module Let_syntax : sig ... end

module Abbrev_type : sig ... end

module Abbreviations : sig ... end

module Solver : sig ... end

val solve : ?debug:Constraints__.Import.Bool.t -> abbrevs:Abbreviations.t -> 'a t -> ('a, [> Solver.error ]) Constraints__.Import.Result.t

solve t solves the constraint t, returning it's value or an error.

type existential_context = Shallow_type.Ctx.t

val sexp_of_existential_context : existential_context -> Sexplib0.Sexp.t

type universal_context = variable Constraints__.Import.List.t

val sexp_of_universal_context : universal_context -> Sexplib0.Sexp.t

type equations = (Type.t * Type.t) Constraints__.Import.List.t

val sexp_of_equations : equations -> Sexplib0.Sexp.t

val (&~) : 'a t -> 'b t -> ('a * 'b) t

&~ is an infix alias for both.

val (>>) : 'a t -> 'b t -> 'b t

t1 >> t2 >> ... >> tn solves t1, ..., tn yielding the value of tn. It is the monodial operator of constraints.

val (=~) : variable -> variable -> Constraints__.Import.Unit.t t

a =~ a' is an infix alias for Eq, denoting the equality constraint on type variables.

val (=~=) : variable -> Shallow_type.t -> Constraints__.Import.Unit.t t

val (=~-) : variable -> Type.t -> Constraints__.Import.Unit.t t

type 'a bound = Decoded.Var.t Constraints__.Import.List.t * 'a

and term_binding = Algebra.Term_var.t * Decoded.scheme

and 'a term_let_binding = term_binding Constraints__.Import.List.t * 'a bound

and 'a term_let_rec_binding = term_binding * 'a bound

val inst : Algebra.Term_var.t -> variable -> Decoded.Type.t Constraints__.Import.List.t t

inst x a is the constraint that instantiates x to a. It returns the type variable substitution.

val decode : variable -> Decoded.Type.t t

decode a is a constraint that evaluates to the decoded type of a.

val exists : ctx:existential_context -> 'a t -> 'a t

exists ~ctx t binds existential context ctx in t.

val forall : ctx:universal_context -> 'a t -> 'a t

forall ~ctx t binds universal context ctx in t.

val (#=) : Algebra.Term_var.t -> variable -> binding

x #= a yields the binding that binds x to a.

val def : bindings:def_binding Constraints__.Import.List.t -> in_:'a t -> 'a t

def ~bindings ~in_ binds bindings in the constraint in_.

val (#=>) : equations -> 'a t -> 'a t

equations #=> t is an alias for the implication constraint Implication (equations, t).

val let_ : bindings:'a let_binding Constraints__.Import.List.t -> in_:'b t -> ('a term_let_binding Constraints__.Import.List.t * 'b) t

let_ ~binding ~in_ binds the let binding binding in the constraint in_.

val let_0 : in_:'a t -> 'a bound t

val let_1 : binding:'a let_binding -> in_:'b t -> ('a term_let_binding * 'b) t

val let_rec : bindings:'a let_rec_binding Constraints__.Import.List.t -> in_:'b t -> ('a term_let_rec_binding Constraints__.Import.List.t * 'b) t

let_rec ~bindings ~in_ recursively binds the let bindings bindings in the constraint in_.

module Binding : sig ... end

Constraints may be used to define "structures" of constraints.

Intuitively these correspond to structures in the expression language.

These structures also form applicatives functors :)

module Structure : sig ... end