Module Constraint.Make

Parameters

module Algebra : Module_types.Algebra_with_decoded

Signature

module Label := Algebra.Types.Label

module Type_var := Algebra.Types.Var

module Type_former := Algebra.Types.Former

variable is the type for constraint "type" variables.

type variable = private Base.int

val sexp_of_variable : variable -> Sexplib0.Sexp.t

fresh () creates a fresh constraint variable.

val fresh : Base.unit -> variable

The module Type provides the concrete representation of types (using constraint type variables) in constraints.

It is the free monad of the functor Type_former.t.

History: This representation was initially used in constraints t, however, the refactor for "Sharing" now uses Shallow_type.t. We however, still use Type for a rich interface.

module Type : sig ... end

The module Shallow_type provides the shallow type definition used within constraints.

This encoding is required for "(Explicit) sharing" of types within constraints.

Types from Type are often referred to as "deep" types.

module Shallow_type : sig ... end

type existential_context = Shallow_type.Ctx.t

val sexp_of_existential_context : existential_context -> Sexplib0.Sexp.t

type universal_context = variable Base.list

val sexp_of_universal_context : universal_context -> Sexplib0.Sexp.t

type equations = (Type.t * Type.t) Base.list

val sexp_of_equations : equations -> Sexplib0.Sexp.t

type 'a bound = Algebra.Decoded.Var.t Base.list * 'a

and term_binding = Algebra.Term_var.t * Algebra.Decoded.scheme

and 'a term_let_binding = term_binding Base.list * 'a bound

and 'a term_let_rec_binding = term_binding * 'a bound

type binding = Algebra.Term_var.t * variable

and def_binding = binding

val sexp_of_binding : binding -> Sexplib0.Sexp.t

val sexp_of_def_binding : def_binding -> Sexplib0.Sexp.t

'a t is a constraint with value type 'a.

In the meta-theory, the constraint language has a defined type system.

In our implementation, we use GADTs to implement the type system, where the type parameter 'a denotes the type of the constraint.

type _ t =

| True : Base.unit t
(*
true
*)

| Return : 'a -> 'a t
(*
return a
*)

| Conj : 'a t * 'b t -> ('a * 'b) t
(*
C₁ && C₂
*)

| Eq : variable * variable -> Base.unit t
(*
ɑ₁ = ɑ₂
*)

| Exist : existential_context * 'a t -> 'a t
(*
exists Θ. C
*)

| Forall : universal_context * 'a t -> 'a t
(*
forall Λ. C
*)

| Instance : Algebra.Term_var.t * variable -> Algebra.Decoded.Type.t Base.list t
(*
x <= ɑ
*)

| Def : binding Base.list * 'a t -> 'a t
(*
def x1 : t1 and ... and xn : tn in C
*)

| Let : 'a let_binding Base.list * 'b t -> ('a term_let_binding Base.list * 'b) t
(*
let Γ in C
*)

| Let_rec : 'a let_rec_binding Base.list * 'b t -> ('a term_let_rec_binding Base.list * 'b) t
(*
let rec Γ in C
*)

| Map : 'a t * ('a -> 'b) -> 'b t
(*
map C f.
*)

| Decode : variable -> Algebra.Decoded.Type.t t
(*
decode ɑ
*)

| Implication : equations * 'a t -> 'a t
(*
E => C
*)

and 'a let_binding =

| Let_binding of {

universal_context : universal_context;

existential_context : existential_context;

is_non_expansive : Base.bool;

bindings : binding Base.list;

in_ : 'a t;

equations : equations;

}

and 'a let_rec_binding =

| Let_rec_mono_binding of {

universal_context : universal_context;

existential_context : existential_context;

binding : binding;

in_ : 'a t;

}

| Let_rec_poly_binding of {

universal_context : universal_context;

annotation : Shallow_type.encoded_type;

term_var : Algebra.Term_var.t;

in_ : 'a t;

}

val sexp_of_t : 'a t -> Base.Sexp.t

'a t forms an applicative functor, allowing us to combine many constraints into a single one using let%map:

val pat : Typedtree.pattern Constraint.t
val exp : Typedtree.expression Constraint.t

let%map pat and exp in
Texp_fun (pat, ..., exp)

include Base.Applicative.S with type 'a t := 'a t

val return : 'a -> 'a t

val map : 'a t -> f:('a -> 'b) -> 'b t

val both : 'a t -> 'b t -> ('a * 'b) t

val (<*>) : ('a -> 'b) t -> 'a t -> 'b t

val (<*) : 'a t -> unit t -> 'a t

val (*>) : unit t -> 'a t -> 'a t

val (>>|) : 'a t -> ('a -> 'b) -> 'b t

val apply : ('a -> 'b) t -> 'a t -> 'b t

val map2 : 'a t -> 'b t -> f:('a -> 'b -> 'c) -> 'c t

val map3 : 'a t -> 'b t -> 'c t -> f:('a -> 'b -> 'c -> 'd) -> 'd t

val all : 'a t list -> 'a list t

val all_unit : unit t list -> unit t

module Applicative_infix : sig ... end

include Base.Applicative.Let_syntax with type 'a t := 'a t

module Open_on_rhs_intf : sig ... end

module Let_syntax : sig ... end

val (&~) : 'a t -> 'b t -> ('a * 'b) t

&~ is an infix alias for both.

val (>>) : 'a t -> 'b t -> 'b t

t1 >> t2 >> ... >> tn solves t1, ..., tn yielding the value of tn. It is the monodial operator of constraints.

val (=~) : variable -> variable -> Base.unit t

a =~ a' is an infix alias for Eq, denoting the equality constraint on type variables.

val (=~=) : variable -> Shallow_type.t -> Base.unit t

val (=~-) : variable -> Type.t -> Base.unit t

val inst : Algebra.Term_var.t -> variable -> Algebra.Decoded.Type.t Base.list t

inst x a is the constraint that instantiates x to a. It returns the type variable substitution.

val decode : variable -> Algebra.Decoded.Type.t t

decode a is a constraint that evaluates to the decoded type of a.

val exists : ctx:existential_context -> 'a t -> 'a t

exists ~ctx t binds existential context ctx in t.

val forall : ctx:universal_context -> 'a t -> 'a t

forall ~ctx t binds universal context ctx in t.

val (#=) : Algebra.Term_var.t -> variable -> binding

x #= a yields the binding that binds x to a.

val def : bindings:def_binding Base.list -> in_:'a t -> 'a t

def ~bindings ~in_ binds bindings in the constraint in_.

val (#=>) : equations -> 'a t -> 'a t

equations #=> t is an alias for the implication constraint Implication (equations, t).

module Binding : sig ... end

val let_ : bindings:'a let_binding Base.list -> in_:'b t -> ('a term_let_binding Base.list * 'b) t

let_ ~binding ~in_ binds the let binding binding in the constraint in_.

val let_0 : in_:'a t -> 'a bound t

val let_1 : binding:'a let_binding -> in_:'b t -> ('a term_let_binding * 'b) t

val let_rec : bindings:'a let_rec_binding Base.list -> in_:'b t -> ('a term_let_rec_binding Base.list * 'b) t

let_rec ~bindings ~in_ recursively binds the let bindings bindings in the constraint in_.

module Structure : sig ... end