Module Make.Type

type t =

| Var of variable

| Former of t Type_former.t

| Row_cons of Label.t * t * t

| Row_uniform of t

| Mu of variable * t

| Let of variable * t * t

t represents the type defined by the grammar: t ::= ɑ | (t, .., t) F

val sexp_of_t : t -> Sexplib0.Sexp.t

val var : variable -> t

var 'a is the representation of the type variable 'a as the type t.

val former : t Type_former.t -> t

former f is the representation of the concrete type former f in type t.

val mu : variable -> t -> t

mu 'a t is the representation of the recursive type mu 'a. t

val let_ : binding:(variable * t) -> in_:t -> t