Module Private.Generalization

Abstract types to be substituted by functor arguments.

Parameters

module Label : sig ... end

module Type_former : Module_types.Type_former.S

Signature

type state

The generalization module maintains several bits of mutable state related to it's implemented of efficient level-based generalization.

We encapsulate this in the abstract type state.

val make_state : Constraints__.Import.Unit.t -> state

make_state () creates a new empty state.

module Rigid_type : sig ... end

module Abbrev_type : sig ... end

module Abbreviations : sig ... end

module Equations : sig ... end

module Structure : sig ... end

module Unifier : sig ... end

val make_rigid_var : state:state -> Constraints__.Import.Int.t -> Unifier.Type.t

make_rigid_var ~state rigid_var creates rigid unification variable.

val make_flexible_var : state:state -> Unifier.Type.t

make_flexible_var ~state creates a flexible unification variable.

val make_former : state:state -> Unifier.Type.t Type_former.t -> Unifier.Type.t

make_former ~state former creates a unification type w/ former former.

val make_row_uniform : state:state -> Unifier.Type.t -> Unifier.Type.t

make_row_uniform ~state type_ creates a unification row w/ row ∂(type_)

val make_row_cons : state:state -> label:Label.t -> field:Unifier.Type.t -> tl:Unifier.Type.t -> Unifier.Type.t

make_row_cons ~state ~label ~field ~tl creates a unification row w/ row (label : field; tl)

The type scheme defines the abstract notion of a scheme in "graphic" types.

A scheme has a notion of "bound" variables, this is encoded in the type variables.

We note that all bound variables to a *generalized* scheme are *rigid*.

type scheme

The notion of "bound" variables for a scheme is shared between instantiation and computing the set of generic variables, given a scheme.

type variables = Unifier.Type.t Constraints__.Import.List.t

val root : scheme -> Unifier.Type.t

root scheme returns the root node, or quantifier "body" of the scheme scheme.

val variables : scheme -> variables

variables scheme computes the bound variables of a scheme scheme.

val mono_scheme : Unifier.Type.t -> scheme

mono_scheme creates a scheme with no generic variables.

enter ~state creates a new "stack frame" in the constraint solver and enters it.

val enter : state:state -> Constraints__.Import.Unit.t

val instantiate : state:state -> scheme -> variables * Unifier.Type.t

instantiate ~state scheme instantates the scheme scheme. It does so, by taking fresh copies of the generic variables, without necessarily copying other bits of the type.

This is designed for efficient sharing of nodes within the "graphic type".

exception Cannot_flexize of Constraints__.Import.Int.t

exception Rigid_variable_escape of Constraints__.Import.Int.t

exception Scope_escape of Unifier.Type.t

val exit : state:state -> rigid_vars:Constraints__.Import.Int.t Constraints__.Import.List.t -> types:Unifier.Type.t Constraints__.Import.List.t -> variables * scheme Constraints__.Import.List.t