Module Generalization.Rigid_type

type t

t represents a rigid type

val sexp_of_t : t -> Sexplib0.Sexp.t

val make_rigid_var : Constraints__.Import.Int.t -> t

make_rigid_var rigid_var returns the rigid type representing the variable rigid_var

val make_former : t Type_former.t -> t

make_former former returns the rigid type representing the rigid type former former.