Module Unifier.Type

type t

t represents a type. See "graphical types".

val sexp_of_t : t -> Sexplib0.Sexp.t

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

val sexp_of_t_hum : t -> Base.Sexp.t

Each graphical type node consists of:

a unique identifier id (used to define a total ordering).

a mutable structure, which contains the node structure.

val id : t -> Base.int

id t returns the unique identifier of the type t.

val structure : t -> t structure

structure t returns the structure of t.

val set_structure : t -> t structure -> Base.unit

set_structure t structure sets the structure of t to structure.

val hash : t -> Base.int

hash t computes the hash of the graphical type t. Based on it's integer field: id.

val create : t structure -> t

make structure creates a new unification type w/ structure structure.

val fold : t -> f:(t -> 'a structure -> 'a) -> var:(t -> 'a) -> mu:(t -> 'a -> 'a) -> 'a

fold t ~f ~var ~mu will perform a fold over the (potentially) cyclic graph defined by the type t.

val to_dot : t -> Base.string

to_dot t returns the graph-viz .dot encoding of t.

val reset_id_cnt : Base.unit -> Base.unit

reset_id_cnt () resets the internal counter used for id.