Module Structure.Rows

Parameters

module Label : sig ... end

module S : S

Signature

type 'a t =

| Structure of 'a S.t

| Row_cons of Label.t * 'a * 'a

| Row_uniform of 'a

type 'a ctx = {

make_var : Constraints__.Import.Unit.t -> 'a;

make_structure : 'a t -> 'a;

super_ : 'a S.ctx;

}

include S with type 'a t := 'a t and type 'a ctx := 'a ctx

val sexp_of_t : ('a -> Sexplib0.Sexp.t) -> 'a t -> Sexplib0.Sexp.t

exception Cannot_merge

Cannot_merge is raised by merge iff the 2 structures are not "consistent".

val merge : ctx:'a ctx -> create:('a t -> 'a) -> unify:('a -> 'a -> Base.unit) -> 'a t -> 'a t -> 'a t

merge ~ctx ~create ~unify t1 t2 determines computes the merged structure of t1 and t2. If the structures are inconsistent, then Cannot_merge is raised.

merge can emit first-order equalities using unify, or create new terms from structures using create.

An additional context ctx is provided since consistency might be contextual.

val map : 'a t -> f:('a -> 'b) -> 'b t

map t ~f traverses t, applying the transformation f.

val iter : 'a t -> f:('a -> Base.unit) -> Base.unit

iter t ~f traverses t, executing f on each child.

val fold : 'a t -> f:('a -> 'b -> 'b) -> init:'b -> 'b

fold t ~f ~init performs the computation of f, traversing over t with the initial accumulator value of init.