Module type Dromedary_lib_unifier.S

Abstract types to be substituted by functor arguments.

type 'a structure

The type 'a structure is the unification structures (with children of type 'a), given by the functor argument Structure.

type 'a ctx

The type ctx is the arbitrary unification context determined by the structure's context, given by Structure.

module Type : sig ... end

unify ~ctx t1 t2 equates the graphical type nodes t1 and t2, and forms a multi-equation node.

Unify (t1, t2) is raised if the two node cannot be unified.

exception Unify of Type.t * Type.t

val unify : ctx:Type.t ctx -> Type.t -> Type.t -> Base.unit