Module Parsing.Parsetree

Parsetree is the abstract syntax tree produced by parsing Dromedary's source code.

type core_type =

| Ptyp_var of string
(*
Type variables 'a.
*)

| Ptyp_arrow of core_type * core_type
(*
Function types T1 -> T2.
*)

| Ptyp_tuple of core_type list
(*
Product (or "tuple") types.
*)

| Ptyp_constr of core_type list * string
(*
A type constructor (or "type former"), name "constr" is used for consistency between OCaml and Dromedary code.
*)

| Ptyp_variant of row
(*
Polymorphic variants [ ... ]
*)

| Ptyp_mu of string * core_type
(*
Equi-recursive types mu 'a. T
*)

| Ptyp_where of core_type * string * core_type
(*
where-binding T where 'a = T
*)

and row = row_field list * closed_flag

and row_field =

| Row_tag of string * core_type option
(*
`A `A of T
*)

and closed_flag =

| Closed
(*
(< | =) row
*)

| Open
(*
> row
*)

val sexp_of_core_type : core_type -> Sexplib0.Sexp.t

val sexp_of_row : row -> Sexplib0.Sexp.t

val sexp_of_row_field : row_field -> Sexplib0.Sexp.t

val sexp_of_closed_flag : closed_flag -> Sexplib0.Sexp.t

val pp_core_type_mach : core_type Fmt.t

pp_core_type_mach ppf core_type pretty prints core_type in a "machine format" (an explicit tree structure).

val pp_core_type : core_type Fmt.t

pp_core_type ppf core_type pretty prints core_type as a syntactic representation.

type core_scheme = string list * core_type

core_schemes are defined by the grammar: core_schemes ::= 'a1 ... 'an. T

This "source"-level notion of type schemes is often used in annotations in expression and pattern.

val sexp_of_core_scheme : core_scheme -> Sexplib0.Sexp.t

val pp_core_scheme_mach : core_scheme Fmt.t

val pp_core_scheme : core_scheme Fmt.t

type pattern =

| Ppat_any
(*
_.
*)

| Ppat_var of string
(*
x.
*)

| Ppat_alias of pattern * string
(*
P as x.
*)

| Ppat_const of Ast_types.constant
(*
c. e.g. 1, true, ().
*)

| Ppat_tuple of pattern list
(*
(P1, ..., Pn). Invariant n >= 2.
*)

| Ppat_construct of string * (string list * pattern) option
(*
C <P>.
*)

| Ppat_variant of string * pattern option
(*
`A <P>.
*)

| Ppat_constraint of pattern * core_type
(*
(P : T).
*)

val sexp_of_pattern : pattern -> Sexplib0.Sexp.t

val pp_pattern_mach : pattern Fmt.t

val pp_pattern : pattern Fmt.t

type expression =

| Pexp_var of string
(*
x.
*)

| Pexp_prim of Ast_types.primitive
(*
Primitive operations %prim p. e.g. %prim +, %prim -, etc.
*)

| Pexp_const of Ast_types.constant
(*
Constants c. e.g. 1, true, ().
*)

| Pexp_fun of pattern * expression
(*
The function (or lambda) abstraction fun P -> E. Note that:

let x P1 ... Pn = E in ... is encoding using Pexp_let ("x", fun P1 ... Pn -> E, ...).

*)

| Pexp_app of expression * expression
(*
Function application E1 E2.
*)

| Pexp_let of Ast_types.rec_flag * value_binding list * expression
(*
Let expressions: let P1 = E1 and ... and Pn = En in E
*)

| Pexp_forall of string list * expression
(*
Explicit "forall" quantifier forall 'a1 ... 'an -> E. Invariant n >= 1.
*)

| Pexp_exists of string list * expression
(*
Explicit "exists" quantifier exists 'a1 ... 'an -> E. Invariant n >= 1.
*)

| Pexp_constraint of expression * core_type
(*
An expression "constraint" or typing annotation (E : T).
*)

| Pexp_construct of string * expression option
(*
An applied algebraic data type constructor C <E>.
*)

| Pexp_record of (string * expression) list
(*
{l1 = E1; ...; ln = En}
*)

| Pexp_field of expression * string
(*
E.l
*)

| Pexp_tuple of expression list
(*
Tuples (E1, ..., En). Invariant: n >= 2.
*)

| Pexp_match of expression * case list
(*
Match (or "case") expressions match E with (P1 -> E1 | ... | Pn -> En).
*)

| Pexp_ifthenelse of expression * expression * expression
(*
If (or ternary) expressions if E then E1 else E2.
*)

| Pexp_try of expression * case list
(*
try E0 with P1 -> E1 | ... | Pn -> En
*)

| Pexp_sequence of expression * expression
(*
E1; E2
*)

| Pexp_while of expression * expression
(*
while E1 do E2 done
*)

| Pexp_for of pattern * expression * expression * Ast_types.direction_flag * expression
(*
for i = E1 to E2 do E3 done for i = E2 downto E2 do E3 done
*)

| Pexp_variant of string * expression option
(*
`A <E>
*)

and value_binding = {

pvb_forall_vars : string list;

pvb_pat : pattern;

pvb_expr : expression;

}

Value bindings P = E

and case = {

pc_lhs : pattern;

pc_rhs : expression;

}

Cases P -> E.

val sexp_of_expression : expression -> Sexplib0.Sexp.t

val sexp_of_value_binding : value_binding -> Sexplib0.Sexp.t

val sexp_of_case : case -> Sexplib0.Sexp.t

val pp_expression_mach : expression Fmt.t

pp_expression_mach ppf exp pretty prints an expression exp as an explicit tree structure.

val pp_expression : expression Fmt.t

pp_expression ppf exp pretty prints an expression exp as a syntax representation.

val pp_value_binding_mach : value_binding Fmt.t

pp_value_binding_mach ppf value_binding pretty prints the value binding value_binding as an explicit tree.

val pp_value_binding : value_binding Fmt.t

pp_value_binding ppf value_binding pretty prints the value binding value_bindings as a syntactic representation.

val pp_case_mach : case Fmt.t

pp_case_mach ppf case pretty prints the case case as an explicit tree structure.

val pp_case : case Fmt.t

pp_case ppf case prety prints the case case as a syntactic representation.

type value_description = {

pval_name : string;

pval_type : core_scheme;

pval_prim : string;

}

External value descriptions external x : T = "prim"

val sexp_of_value_description : value_description -> Sexplib0.Sexp.t

val pp_value_description_mach : value_description Fmt.t

pp_value_description_mach ppf value_desc pretty prints value description value_desc as an explicit tree structure.

type type_declaration = {

ptype_name : string;

ptype_params : string list;
(*
Parameters: ('a1, .., 'an) t
*)

ptype_kind : type_decl_kind;
(*
Declaration kinds: record or variant
*)

}

and type_decl_kind =

| Ptype_variant of constructor_declaration list

| Ptype_record of label_declaration list

| Ptype_abstract

| Ptype_alias of core_type

| Ptype_open

and label_declaration = {

plabel_name : string;
(*
Label name l
*)

plabel_betas : string list;
(*
Label betas, used for semi-explicit first-class polymorphism
*)

plabel_arg : core_type;
(*
Label argument type
*)

}

Label declaration: l : 'b1 .. 'bm. T

and constructor_declaration = {

pconstructor_name : string;
(*
Constructor name C
*)

pconstructor_arg : constructor_argument option;
(*
Constructor argument
*)

pconstructor_constraints : (core_type * core_type) list;
(*
Constructor constraints (used for GADTs)
*)

}

Constructor declaration: C of 'b1 .. 'bm. T constraint E

and constructor_argument = {

pconstructor_arg_betas : string list;

pconstructor_arg_type : core_type;

}

val sexp_of_type_declaration : type_declaration -> Sexplib0.Sexp.t

val sexp_of_type_decl_kind : type_decl_kind -> Sexplib0.Sexp.t

val sexp_of_label_declaration : label_declaration -> Sexplib0.Sexp.t

val sexp_of_constructor_declaration : constructor_declaration -> Sexplib0.Sexp.t

val sexp_of_constructor_argument : constructor_argument -> Sexplib0.Sexp.t

val pp_type_declaration_mach : type_declaration Fmt.t

pp_type_declaration_mach ppf type_decl pretty prints type_decl using explicit tree structure.

val pp_type_declaration : type_declaration Fmt.t

pp_type_declaration ppf type_decl pretty prints a type_decl as a syntax representation.

val pp_label_declaration_mach : label_declaration Fmt.t

pp_label_declaration_mach ppf label_decl pretty prints label_decl using explicit tree structure.

val pp_label_declaration : label_declaration Fmt.t

pp_label_declaration ppf label_decl pretty prints a label_decl as a syntax representation.

val pp_constructor_declaration_mach : constructor_declaration Fmt.t

pp_constructor_declaration_mach ppf constr_decl pretty prints constr_decl using explicit tree structure.

val pp_constructor_declaration : constructor_declaration Fmt.t

pp_constructor_declaration ppf constr_decl pretty prints a constr_decl as a syntax representation.

type extension_constructor = {

pext_kind : extension_constructor_kind;

}

Extension constructor type ('a1, ... 'an) t += ...

and extension_constructor_kind =

| Pext_decl of constructor_declaration
(*
C of 'b1 .. 'bm. T constraint E
*)

val pp_extension_constructor_mach : extension_constructor Fmt.t

pp_extension_constructor_mach ppf ext_constr pretty prints ext_constr using explicit tree structure.

val pp_extension_constructor : extension_constructor Fmt.t

pp_extension_constructor ppf ext_constr pretty prints a ext_constr as a syntax representation.

type type_exception = {

ptyexn_constructor : extension_constructor;

}

Exception exception C <of T>

type type_extension = {

ptyext_name : string;

ptyext_params : string list;
(*
Parameters: ('a1, .., 'an) t
*)

ptyext_constructors : extension_constructor list;
(*
Constructors C1 [of t1] | ... | Cn [of tn]
*)

}

Extensions type t += ..

val sexp_of_type_extension : type_extension -> Sexplib0.Sexp.t

type structure_item =

| Pstr_value of Ast_types.rec_flag * value_binding list
(*
Structure let binding: let <rec> P1 = E1 and ... and Pn = En
*)

| Pstr_primitive of value_description
(*
External primitive descriptions
*)

| Pstr_type of type_declaration list
(*
Type declarations type t1 = ... and ... tn = ...
*)

| Pstr_typext of type_extension
(*
Type extensions type t += ...
*)

| Pstr_exception of type_exception
(*
Exception exception C <of T>
*)

val sexp_of_structure_item : structure_item -> Sexplib0.Sexp.t

val pp_structure_item_mach : structure_item Fmt.t

val pp_structure_item : structure_item Fmt.t

type structure = structure_item list

Structures -- a list of structure items

val sexp_of_structure : structure -> Sexplib0.Sexp.t

val pp_structure_mach : structure Fmt.t

val pp_structure : structure Fmt.t