Module Private.Algebra

module Term_var : sig ... end

Term_var implements the abstract notion of variables in Dromedary's expressions (or "Terms").

module Type_var : sig ... end

Type_var implements the abstract notion of type variables in Dromedary's types.

module Type_former : sig ... end

Type_former defines the various type formers for Dromedary's types.

include Constraints.Module_types.Algebra with module Term_var := Term_var and type Types.Label.t = string and module Types.Var = Type_var and module Types.Former = Type_former

module Types : sig ... end