Module Computation.Pattern

include S

type 'a t

A computation 'a Computation.t represents a monadic computation that produces a value of type 'a.

Computations are designed for computating (or generating) 'a Constraint.t's, thus it's syntax provided by ppx_let is altered (from standard Monadic Let_syntax) for this.

Computations are bound using the let%bind syntax:

val comp1 : Typedtree.pattern Constraint.t Computation.t

let%bind pat1 = comp1 in
let%bind pat2 = comp1 in
...

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

val (>>|) : 'a t -> ('a -> 'b) -> 'b t

module Monad_infix : sig ... end

val bind : 'a t -> f:('a -> 'b t) -> 'b t

val return : 'a -> 'a t

val map : 'a t -> f:('a -> 'b) -> 'b t

val join : 'a t t -> 'a t

val ignore_m : 'a t -> unit t

val all : 'a t list -> 'a list t

val all_unit : unit t list -> unit t

val const : 'a -> 'a Constraints.Make(Typing__.Import.Algebra).t t

const x creates a computation that returns a constraint 'a Constraint.t that evaluates to x.

val fail : Typing__.Import.Sexp.t -> 'a t

fail err raises the error err.

val env : Env.t t

env returns the environment

val find_label : string -> Types.label_declaration t

find_label label returns the corresponding label declaration in the environment w/ label label.

val find_constr : string -> Types.constructor_declaration t

find_constr name returns the corresponding constructor declaration in the environment w/ constructor name name.

val substitution : Typing__.Substitution.t t

substitution returns the local substitution.

val extend_substitution : 'a t -> substitution:Typing__.Substitution.t -> 'a t

val extend_substitution_vars : in_:(Constraints.Make(Typing__.Import.Algebra).variable list -> 'a t) -> vars:string list -> on_duplicate_var:(string -> Typing__.Import.Sexp.t) -> 'a t

val find_var : string -> Constraints.Make(Typing__.Import.Algebra).variable t

find_var var returns the corresponding constraint variable for the bound type variable var in the local substitution.

val of_result : ('a, 'err) Typing__.Import.Result.t -> message:('err -> Typing__.Import.Sexp.t) -> 'a t

of_result result ~message lifts the result result into a computation, using message to compute the error message for the computation.

module Binder : sig ... end

Let_syntax does not follow the conventional Let_syntax signature for a Monad. Instead we have standard return and bind, however, the map and both are used for constructing constraints.

This allows the pattern for constructing constraints:

let%bind p1 = comp1 in
let%bind p2 = comp2 in
return
  (let%map () = var1 =~ var2 in
   ...)

Binders are bound using the let-op let@.

module Let_syntax : sig ... end

val write : Fragment.t -> unit t

val extend : string -> Constraints.Make(Typing__.Import.Algebra).variable -> unit t

val assert_ : (Constraints.Make(Typing__.Import.Algebra).Type.t * Constraints.Make(Typing__.Import.Algebra).Type.t) Constraints__.Import.List.t -> unit t

val extend_fragment_substitution : Typing__.Substitution.t -> unit t

val run : 'a t -> (Fragment.t * 'a) Expression.t