Module Computation.Binder

type 'a computation := 'a t

type 'a t

A 'a Binder.t represents a monadic binding context for a 'b Constraint.t Computation.t.

They are designed to provide an intuitive notion of "compositional" binding (avoiding continuation hell!).

Computations are bound using the let-op let&.

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

val (>>|) : 'a t -> ('a -> 'b) -> 'b t

module Monad_infix : sig ... end

val bind : 'a t -> f:('a -> 'b t) -> 'b t

val return : 'a -> 'a t

val map : 'a t -> f:('a -> 'b) -> 'b t

val join : 'a t t -> 'a t

val ignore_m : 'a t -> unit t

val all : 'a t list -> 'a list t

val all_unit : unit t list -> unit t

val exists : unit -> Constraints.Make(Typing__.Import.Algebra).variable t

val forall : unit -> Constraints.Make(Typing__.Import.Algebra).variable t

val exists_vars : Constraints.Make(Typing__.Import.Algebra).variable list -> unit t

val forall_ctx : ctx: Constraints.Make(Typing__.Import.Algebra).variable Constraints__.Import.List.t -> unit t

val exists_ctx : ctx: (Constraints.Make(Typing__.Import.Algebra).variable Constraints__.Import.List.t * (Constraints.Make(Typing__.Import.Algebra).variable * Constraints.Make(Typing__.Import.Algebra).Shallow_type.t) Constraints__.Import.List.t) -> unit t

val of_type : Constraints.Make(Typing__.Import.Algebra).Type.t -> Constraints.Make(Typing__.Import.Algebra).variable t

module Let_syntax : sig ... end