Module Typing.Types

Representation of types and declarations

type tag = string

module Type_var : sig ... end

module Type_expr : sig ... end

type type_expr = Type_expr.t

val sexp_of_type_expr : type_expr -> Sexplib0.Sexp.t

type row = Type_expr.t

val sexp_of_row : row -> Sexplib0.Sexp.t

type type_var = Type_var.t

type scheme = type_var list * type_expr

val sexp_of_scheme : scheme -> Sexplib0.Sexp.t

val sexp_of_type_var : type_var -> Typing__.Import.Sexp.t

val pp_type_expr_mach : indent:string -> type_expr Fmt.t

pp_type_expr_mach ppf type_expr pretty prints an explicit tree of the type expression type_expr.

val pp_type_expr : type_expr Fmt.t

pp_type_expr ppf type_expr pretty prints the syntactic representation of the type expression type_expr.

type type_declaration = {

type_name : string;

type_kind : type_decl_kind;

}

and type_decl_kind =

| Type_record of label_declaration list

| Type_variant of constructor_declaration list

| Type_abstract

| Type_alias of alias

| Type_open

and alias = {

alias_alphas : type_var list;

alias_name : string;

alias_type : type_expr;

}

and label_declaration = {

label_name : string;

label_alphas : type_var list;

label_betas : type_var list;

label_arg : type_expr;

label_type : type_expr;

}

and constructor_declaration = {

constructor_name : string;

constructor_alphas : type_var list;

constructor_arg : constructor_argument option;

constructor_type : type_expr;

constructor_constraints : (type_expr * type_expr) list;

}

and constructor_argument = {

constructor_arg_betas : type_var list;

constructor_arg_type : type_expr;

}

val sexp_of_type_declaration : type_declaration -> Sexplib0.Sexp.t

val sexp_of_type_decl_kind : type_decl_kind -> Sexplib0.Sexp.t

val sexp_of_alias : alias -> Sexplib0.Sexp.t

val sexp_of_label_declaration : label_declaration -> Sexplib0.Sexp.t

val sexp_of_constructor_declaration : constructor_declaration -> Sexplib0.Sexp.t

val sexp_of_constructor_argument : constructor_argument -> Sexplib0.Sexp.t

val pp_constructor_declaration_mach : indent:string -> constructor_declaration Fmt.t

val pp_type_declaration_mach : indent:string -> type_declaration Fmt.t

type constructor_description = {

constructor_name : string;

constructor_arg : type_expr option;

constructor_type : type_expr;

}

val sexp_of_constructor_description : constructor_description -> Sexplib0.Sexp.t

val pp_constructor_description_mach : indent:string -> constructor_description Fmt.t

val pp_constructor_description : constructor_description Fmt.t

type variant_description = {

variant_tag : tag;

variant_row : row;

}

val sexp_of_variant_description : variant_description -> Sexplib0.Sexp.t

val pp_variant_description_mach : indent:string -> variant_description Fmt.t

val pp_variant_description : variant_description Fmt.t

type label_description = {

label_name : string;

label_arg : type_expr;

label_type : type_expr;

}

val sexp_of_label_description : label_description -> Sexplib0.Sexp.t

val pp_label_description_mach : indent:string -> label_description Fmt.t

val pp_label_description : label_description Fmt.t